Increases test coverage some more, where possible
PS. At the moment, bastard injection on some static ambient REST actions
is the lesser of two evils (the other being not unit testing the method
relying on it). Since later on we want to perhaps revisit the REST stuff
anyways, we might as well go down this path for now.