Notifiers'll get details for own NotificationLevel
When a Notifier has a threshold of (e.g.) "Warning", then all IChecks
that gave a warning should also render their detailed message to that
notifier. So the Notifiers now pass this level down to the chain of
RenderPlainText calls.

The parameterless version of RenderPlainText was left in this commit
intentionally, for two related reasons:

1. There might be value in keeping it (even though it might often be
   unexpected that it only gives back Error level messages).
2. Even if it is removed, it's better to do so in a separate commit so
   that it can be separately reviewed, reverted, etc.

Fixes #3