Removes parameterless RenderPlainText methods
The risk of someone writing an INotifier implementation and not getting
check result details because the parameterless method only returns de-
tails for Error level results is real.

We *might* need to re-introduce the method later, but will only do so
when it makes sense and is needed. Until then, it's probably API bloat
that will confuse authors.