Sets up NuGet packaging stuff
First "just good enough" state for versioning the assemblies and making
a NuGet package out of the source code. There's a ton of things to
improve on this set-up, as it requires quite a bit of manual labor to
get the versions and packaging just right. But it'll do for now.